requirements.txt
catboost==1.0.6
joblib==1.1.0
numpy==1.21.5
pandas==1.4.4
Pillow==9.4.0
scikit_learn==1.2.0
streamlit==1.11.0
xgboost==1.6.1
